refactor(ignore): filter abstruction
